Move Embeddings and CircuitSchedule
