+++ qiskit_ibm_runtime/utils/__init__.py
@@ -18,7 +18,7 @@ from .converters import (
     seconds_to_duration,
     duration_difference,
 )
-from .embeddings import Embedding
+from ..visualization.embeddings import Embedding
 from .utils import (
     to_python_identifier,
     is_crn,
