+++ qiskit_ibm_runtime/visualization/draw_circuit_schedule_timings.py
@@ -16,7 +16,7 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
-from ..utils.circuit_schedule import CircuitSchedule
+from .circuit_schedule import CircuitSchedule
 from .utils import plotly_module
 
 if TYPE_CHECKING:
